+++ src/unpoly/layer.js
@@ -1243,7 +1243,7 @@ up.layer = (function() {
   The *current* layer is usually the [frontmost layer](/up.layer.front).
   There are however some cases where the current layer is a layer in the background:
 
-  - While an element in a background layer is being [compiled](/up.compiler).
+  - While an element in a background layer is being [compiled](/enhancing-elements).
   - While an Unpoly event like `up:request:loaded` is being triggered from a background layer.
   - While an event listener bound to a background layer using `up.Layer#on()` is being called.